New data fields in Item Schema
add the author, likes, photo,and numSessions (number of times exercise app user community have done a workout). PhotoSchema setup will be decided after talking to the group. Considering defining photoschema within the ItemSchema to organize the code better, and to make it easy to set default values to the image.